requirements.txt
torch==2.2.0 
torchvision==0.17.0
gradio>=4.0.0
Pillow>=9.0.0
numpy==1.26.4
basicsr @ git+https://github.com/XPixelGroup/BasicSR@8d56e3a045f9fb3e1d8872f92ee4a4f07f886b0a
realesrgan>=0.3.0
opencv-python>=4.5.0
